Update sign up page
