+++ twitter/twitter/twitter.py
@@ -1,7 +1,7 @@
 """Welcome to Pynecone! This file outlines the steps to create a basic app."""
 import pynecone as pc
 from .base_state import State
-from .auth import login, signup
+from .pages import login, signup
 from .home import home
 
 app = pc.App(state=State)
